Some progress I guess

+++ src/main/java/cn/nukkit/network/protocol/InventorySlotPacket.java
@@ -30,7 +30,7 @@ public class InventorySlotPacket extends DataPacket {
     @Override
     public void encode() {
         this.reset();
-        this.putUnsignedVarInt((byte) this.inventoryId);
+        this.putUnsignedVarInt(this.inventoryId);
         this.putUnsignedVarInt(this.slot);
         if (protocol >= 407) {
             this.putVarInt(networkId);
